ai-interviewer/backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
opencv-python==4.5.3.56
numpy==1.21.2
pymongo==3.12.0
google-generativeai==0.3.0
SpeechRecognition==3.8.1
pyttsx3==2.90
python-dotenv==0.19.0
mediapipe==0.8.7.3
scikit-learn==0.24.2
textstat==0.7.2
fuzzywuzzy==0.18.0